Start of refactor for moving imageing to a package rather than a main
