+++ imagetocsv/main.go
@@ -1,4 +1,4 @@
-package main
+package imagetocsv
 
 import (
 	"fmt"
@@ -41,26 +41,4 @@ func imageToCSV(filename string) string {
 		}
 	}
 	return strings.Join(vals, ",")
-}
-
-func main() {
-	if (len(os.Args) < 2) {
-		fmt.Printf("Usage: imagetocsv out_file in_file(s)")
-		return
-	}
-
-	args := os.Args[1:]
-	outfilename := args[0]
-
-	// open output file
-	outfile, err := os.Create(outfilename)
-	if err != nil {
-		panic(err)
-	}
-	defer outfile.Close()
-
-	// write csvs
-	for _, arg := range args[1:] {
-		outfile.WriteString(imageToCSV(arg))
-	}
 }
