--- imaging/imagePros.go
@@ -1,14 +1,0 @@
-package imaging
-
-import (
-    "image"
-    "io"
-
-_   "image/png"
-_   "image/jpeg"
-_   "image/gif"
-)
-
-func convertImage (r Reader) {
-
-}
